+++ extensions/omni_kube/CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.2] - 2025-09-17
+
+### Changed
+
+* `load_kubeconfig` is now `security definer` [#960](https://github.com/omnigres/omnigres/pull/960)
+
+### Fixed
+
+* Internal setting permission error [#960](https://github.com/omnigres/omnigres/pull/960)
+
 ## [0.4.1] - 2025-09-16
 
 ### Fixed
@@ -80,3 +90,5 @@ Initial release
 [0.4.0]: [https://github.com/omnigres/omnigres/pull/948]
 
 [0.4.1]: [https://github.com/omnigres/omnigres/pull/959]
+
+[0.4.2]: [https://github.com/omnigres/omnigres/pull/960]